Restructure project to be aligned with clean architecture
