+++ app/src/main/java/com/example/testwithpoetry/presentation/main/MainViewModel.kt
@@ -1,14 +1,16 @@
-package com.example.testwithpoetry
+package com.example.testwithpoetry.presentation.main
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.testwithpoetry.data.network.NetworkResource
+import com.example.testwithpoetry.data.repository.PoetryRepositoryImpl
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class MainViewModel @Inject constructor(
-    private val repo: PoetryRepository
+    private val repo: PoetryRepositoryImpl
 ): ViewModel() {
     fun action() {
         viewModelScope.launch {
